Styling Part 3 - Home
Before i start the responsiveness to mobile, i had to change some of the Home styles, i wasn't satisfied with how it was looking, going to go through projects and projectPage to change the layout if need be.

Home:
Removed Courses Component.
Intro and ImageIntro has been transformed into components,