Styling - Part 4 - Home  Part 2
App:
installed:
include-media: 1.4.10 (include-media is a Sass library for writing CSS media queries in an easy and maintainable way, using a natural and simplistic syntax.)

Started adapting the Home page to be responsive to different interfaces. For that I'll be using popular breakpoints for different devices. Starting with phones. (<=425px)